# server/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DoneList API
  description: Task management API with category and tag support
  version: 1.0.0
  contact:
    name: API Support
    email: support@donelist.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.donelist.com/v1
    description: Production server
security:
  - BearerAuth: []
...
paths:
...
  /calendar/monthly:
    get:
      tags:
        - Calendar
      summary: Get monthly calendar view
      description: |
        Returns a complete monthly calendar view with daily summaries, streaks, and statistics.
        The response includes weeks formatted for UI rendering, with support for both Sunday and Monday week starts.

        **Rate Limit:** 100 requests per minute
      operationId: getMonthlyCalendar
      parameters:
        - name: year
          in: query
          description: Year (1970-2100)
          schema:
            type: integer
            minimum: 1970
            maximum: 2100
            example: 2024
        - name: month
          in: query
          description: Month (1-12)
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 11
        - name: start_day
          in: query
          description: First day of the week for calendar display
          schema:
            type: string
            enum: [sunday, monday]
            default: monday
        - name: timezone
          in: query
          description: IANA timezone identifier
          schema:
            type: string
            default: UTC
            example: America/New_York
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Monthly calendar data
          headers:
            Cache-Control:
              description: Cache control directives
              schema:
                type: string
                example: 'public, max-age=3600'
            Expires:
              description: Cache expiration time
              schema:
                type: string
                format: http-date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyCalendar'
              examples:
                november_2024:
                  summary: November 2024 calendar with activity
                  value:
                    year: 2024
                    month: 11
                    month_name: November
                    start_day: monday
                    weeks:
                      - days:
                        - date: '2024-10-28'
                          checkin_count: 0
                          total_minutes: 0
                          completion_percent: 0
                          color_intensity: 0
                          is_current_month: false
                          is_today: false
                          has_checkins: false
                    summary:
                      total_checkins: 50
                      total_minutes: 1200
                      days_with_checkins: 20
                      total_days_in_month: 30
                      average_per_day: 1.67
                      completion_rate: 66.67
                      most_productive_day: '2024-11-15'
                      most_productive_count: 8
                      current_streak: 5
                      longest_streak: 12
                    categories:
                      - category_id: '550e8400-e29b-41d4-a716-446655440000'
                        category_name: Work
                        count: 30
                        percentage: 60
                    previous_month: '2024-10'
                    next_month: '2024-12'
                    generated_at: '2024-11-15T10:30:00Z'
                    cache_expiration: '2024-11-15T11:30:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /calendar/month:
    get:
      tags:
        - Calendar
      summary: Get monthly calendar view (v2)
      description: |
        Returns a monthly calendar view with enhanced caching support via ETags.

        **Features:**
        - Schema versioning for API stability
        - ETag-based conditional requests (304 Not Modified)
        - 15-minute cache TTL

        **Rate Limit:** 100 requests per minute
      operationId: getMonthlyCalendarV2
      parameters:
        - name: year
          in: query
          required: true
          description: Year (1970-2100)
          schema:
            type: integer
            minimum: 1970
            maximum: 2100
            example: 2024
        - name: month
          in: query
          required: true
          description: Month (1-12)
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 11
        - name: startOfWeek
          in: query
          description: First day of the week (sun or mon)
          schema:
            type: string
            enum: [sun, mon]
            default: mon
        - name: timezone
          in: query
          description: IANA timezone identifier
          schema:
            type: string
            default: UTC
            example: America/New_York
        - name: If-None-Match
          in: header
          description: ETag from previous response for conditional requests
          schema:
            type: string
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Monthly calendar data with versioning
          headers:
            ETag:
              description: Entity tag for caching
              schema:
                type: string
                example: '"a1b2c3d4e5f6"'
            Cache-Control:
              description: Cache control directives
              schema:
                type: string
                example: 'private, max-age=900'
            X-Schema-Version:
              description: Response schema version
              schema:
                type: string
                example: '1.0.0'
            Expires:
              description: Cache expiration time
              schema:
                type: string
                format: http-date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyCalendarV2Response'
        '304':
          description: Not Modified - Use cached version when ETag matches
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /calendar/heatmap:
    get:
      tags:
        - Calendar
      summary: Get heatmap data
      description: |
        Returns daily activity heatmap data for visualization (GitHub-style contribution graph).

        **Constraints:**
        - Maximum date range: 366 days
        - Recommended for year-at-a-glance views

        **Rate Limit:** 100 requests per minute
      operationId: getHeatmap
      parameters:
        - name: start_date
          in: query
          required: true
          description: Start date in YYYY-MM-DD format
          schema:
            type: string
            format: date
            example: '2024-01-01'
        - name: end_date
          in: query
          required: true
          description: End date in YYYY-MM-DD format
          schema:
            type: string
            format: date
            example: '2024-12-31'
        - name: timezone
          in: query
          description: IANA timezone identifier
          schema:
            type: string
            default: UTC
            example: America/New_York
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Heatmap data
          headers:
            Cache-Control:
              description: Cache control directives
              schema:
                type: string
                example: 'public, max-age=3600'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeatmapData'
              examples:
                year_2024:
                  summary: Full year 2024 heatmap
                  value:
                    start_date: '2024-01-01'
                    end_date: '2024-12-31'
                    days:
                      - date: '2024-01-01'
                        checkin_count: 5
                        total_minutes: 120
                        completion_percent: 8.33
                        color_intensity: 1
                    total_days: 366
                    active_days: 250
                    total_checkins: 1500
                    generated_at: '2024-11-15T10:30:00Z'
        '400':
          description: Bad request - Invalid date format or range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_format:
                  value:
                    error: validation_error
                    message: invalid start_date format, use YYYY-MM-DD
                range_too_large:
                  value:
                    error: validation_error
                    message: date range cannot exceed 366 days
                invalid_range:
                  value:
                    error: validation_error
                    message: end_date must be after start_date
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    XRequestId:
      name: X-Request-ID
      in: header
      description: Request ID for tracing
      schema:
        type: string
        format: uuid
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        request_id:
          type: string
          format: uuid
          description: Request ID for tracing

    # Timeline Schemas
...
    MonthlyCalendar:
      type: object
      required:
        - year
        - month
        - month_name
        - start_day
        - weeks
        - summary
        - previous_month
        - next_month
        - generated_at
      properties:
        year:
          type: integer
          minimum: 1970
          maximum: 2100
          description: Calendar year
          example: 2024
        month:
          type: integer
          minimum: 1
          maximum: 12
          description: Calendar month (1-12)
          example: 11
        month_name:
          type: string
          description: Full month name
          example: November
        start_day:
          type: string
          enum: [sunday, monday]
          description: First day of the week
          example: monday
        weeks:
          type: array
          description: Calendar weeks (4-6 weeks depending on month)
          minItems: 4
          maxItems: 6
          items:
            $ref: '#/components/schemas/CalendarWeek'
        summary:
          $ref: '#/components/schemas/MonthlySummary'
        categories:
          type: array
          description: Check-in breakdown by category
          items:
            $ref: '#/components/schemas/CategoryBreakdown'
        previous_month:
          type: string
          pattern: '^\d{4}-\d{2}$'
          description: Previous month in YYYY-MM format
          example: '2024-10'
        next_month:
          type: string
          pattern: '^\d{4}-\d{2}$'
          description: Next month in YYYY-MM format
          example: '2024-12'
        generated_at:
          type: string
          format: date-time
          description: Timestamp when calendar was generated
        cache_expiration:
          type: string
          format: date-time
          description: When cached data expires
    CalendarWeek:
      type: object
      required:
        - days
      properties:
        days:
          type: array
          description: Seven days in the week
          minItems: 7
          maxItems: 7
          items:
            $ref: '#/components/schemas/CalendarDay'
...
    MonthlySummary:
      type: object
      required:
        - total_checkins
        - total_minutes
        - days_with_checkins
        - total_days_in_month
        - average_per_day
        - completion_rate
      properties:
        total_checkins:
          type: integer
          minimum: 0
          description: Total number of check-ins in the month
          example: 50
        total_minutes:
          type: integer
          minimum: 0
          description: Total duration of all check-ins in minutes
          example: 1200
        days_with_checkins:
          type: integer
          minimum: 0
          description: Number of days with at least one check-in
          example: 20
        total_days_in_month:
          type: integer
          minimum: 28
          maximum: 31
          description: Total days in the month
          example: 30
        average_per_day:
          type: number
          format: float
          minimum: 0
          description: Average check-ins per day
          example: 1.67
        completion_rate:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Percentage of days with check-ins
          example: 66.67
        most_productive_day:
          type: string
          format: date
          description: Date with most check-ins
          example: '2024-11-15'
        most_productive_count:
          type: integer
          minimum: 0
          description: Check-in count on most productive day
          example: 8
        current_streak:
          type: integer
          minimum: 0
          description: Current consecutive days with check-ins
          example: 5
        longest_streak:
          type: integer
          minimum: 0
          description: Longest streak in this month
          example: 12
    CategoryBreakdown:
      type: object
      required:
        - category_name
        - count
        - percentage
      properties:
        category_id:
          type: string
          format: uuid
          description: Category unique identifier (null for uncategorized)
        category_name:
          type: string
          description: Category name
          example: Work
        count:
          type: integer
          minimum: 0
          description: Number of check-ins in this category
          example: 30
        percentage:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Percentage of total check-ins
          example: 60
    MonthlyCalendarV2Response:
      type: object
      required:
        - schemaVersion
        - data
      properties:
        schemaVersion:
          type: string
          pattern: '^\d+\.\d+\.\d+$'
          description: API response schema version (semver)
          example: '1.0.0'
        data:
          $ref: '#/components/schemas/MonthlyCalendar'
    HeatmapData:
      type: object
      required:
        - start_date
        - end_date
        - days
        - total_days
        - active_days
        - total_checkins
        - generated_at
      properties:
        start_date:
          type: string
          format: date
          description: Start date of the range
          example: '2024-01-01'
        end_date:
          type: string
          format: date
          description: End date of the range
          example: '2024-12-31'
        days:
          type: array
          description: Array of days with activity data
          items:
            $ref: '#/components/schemas/HeatmapDay'
        total_days:
          type: integer
          minimum: 1
          maximum: 366
          description: Total number of days in the range
          example: 366
        active_days:
          type: integer
          minimum: 0
          description: Days with at least one check-in
          example: 250
        total_checkins:
          type: integer
          minimum: 0
          description: Total check-ins in the range
          example: 1500
        generated_at:
          type: string
          format: date-time
          description: Timestamp when data was generated
    HeatmapDay:
      type: object
      required:
        - date
        - checkin_count
        - total_minutes
        - completion_percent
        - color_intensity
      properties:
        date:
          type: string
          format: date
          description: Date in YYYY-MM-DD format
          example: '2024-01-15'
        checkin_count:
          type: integer
          minimum: 0
          description: Number of check-ins
          example: 5
        total_minutes:
          type: integer
          minimum: 0
          description: Total duration in minutes
          example: 120
        completion_percent:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Percentage of day covered
          example: 8.33
        color_intensity:
          type: integer
          minimum: 0
          maximum: 4
          description: Color intensity level (0-4)
          example: 1
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation_error:
              value:
                error: validation_error
                message: Invalid input data
                details:
                  field: name
                  reason: Name must be between 1 and 50 characters
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid_token:
              value:
                error: unauthorized
                message: Invalid or expired authentication token
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            server_error:
              value:
                error: internal_server_error
                message: An unexpected error occurred